refactor(webapp): extract and guard cap-axis helpers
Move niceStep/niceCapAxis out of BreakdownChart into chartFormat so
they can be unit tested without pulling recharts into a test, and
guard niceStep against non-positive input (log10(0) would yield NaN).

Memoize isSeriesHidden in useChartInteractions so BreakdownChart's
cap-axis useMemo only recomputes when the visible series set changes,
not on every render. Also clarify the renderAsArea and cumulative
comments, and rename viewToggle to viewToggleControl so it reads as a
node rather than an action.
